Reorganize combinations code
