+++ src/pykeen/models/multimodal/distmult_literal.py
@@ -7,10 +7,10 @@ from typing import Any, ClassVar, Mapping, Optional
 import torch.nn as nn
 
 from .base import LiteralModel
-from .combinations import DistMultCombination
 from ...constants import DEFAULT_DROPOUT_HPO_RANGE, DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
 from ...losses import Loss
 from ...nn import EmbeddingSpecification
+from ...nn.combinations import DistMultCombination
 from ...nn.modules import DistMultInteraction, LiteralInteraction
 from ...regularizers import Regularizer
 from ...triples import TriplesNumericLiteralsFactory
